refactor: 거래추가 mutation
