+++ src/types/dtos/transaction.ts
@@ -1,0 +1,15 @@
+export type AccountTypeDTO = {
+  typeName: 'ASSET' | 'LIABILITY' | 'EQUITY' | 'REVENUE' | 'EXPENSE';
+  typeState: 'INCREASE' | 'DECREASE' | 'OCCURRENCE';
+};
+
+export type TransactionDTO = {
+  date: Date;
+  summary: string | null;
+  accounts: {
+    accountType: AccountTypeDTO;
+    amount: number;
+    tertiaryCategoryId: number;
+    cardId?: number;
+  }[];
+};
